Deduplicate test fixture rmock into conftest files
