+++ tests/test_data_api_client.py
@@ -1,28 +1,11 @@
 # -*- coding: utf-8 -*-
 from flask import json, request
-import requests_mock
 import pytest
 import mock
 
 from dmapiclient import DataAPIClient
 from dmapiclient import APIError, HTTPError
 from dmapiclient.audit import AuditTypes
-
-
-@pytest.yield_fixture
-def rmock():
-    with requests_mock.mock() as rmock:
-        real_register_uri = rmock.register_uri
-
-        def register_uri_with_complete_qs(*args, **kwargs):
-            if 'complete_qs' not in kwargs:
-                kwargs['complete_qs'] = True
-
-            return real_register_uri(*args, **kwargs)
-
-        rmock.register_uri = register_uri_with_complete_qs
-
-        yield rmock
 
 
 @pytest.fixture
