+++ tests/test_search_api_client.py
@@ -2,28 +2,11 @@
 import os
 
 from flask import json
-import requests_mock
 import pytest
 import mock
 
 from dmapiclient import SearchAPIClient
 from dmapiclient import APIError, HTTPError
-
-
-@pytest.yield_fixture
-def rmock():
-    with requests_mock.mock() as rmock:
-        real_register_uri = rmock.register_uri
-
-        def register_uri_with_complete_qs(*args, **kwargs):
-            if 'complete_qs' not in kwargs:
-                kwargs['complete_qs'] = True
-
-            return real_register_uri(*args, **kwargs)
-
-        rmock.register_uri = register_uri_with_complete_qs
-
-        yield rmock
 
 
 @pytest.fixture
